price.sql, latest_price.sql: Use BIGINT for fixed-point values
To be compatible with PostgreSQL (still works with SQLite).

+++ price.sql
@@ -1,9 +1,9 @@
 CREATE TABLE price(
 security VARCHAR(36) NOT NULL REFERENCES security(uuid),
 tstamp VARCHAR(32) NOT NULL,
-value INT NOT NULL,
-high INT,
-low INT,
-volume INT
+value BIGINT NOT NULL,
+high BIGINT,
+low BIGINT,
+volume BIGINT
 );
 CREATE UNIQUE INDEX price__security_tstamp ON price(security, tstamp);
